+++ hs-bindgen/test/hs-bindgen/Test/HsBindgen/Unit/Frontend.hs
@@ -128,7 +128,8 @@ execFrontend getTestResources cStdStr incDirs header k =
           config         = BindgenConfig bootConfig frontendConfig backendConfig
           bootTracer     = contramap TraceBoot tracer
           frontendTracer = contramap TraceFrontend tracer
-      bootArtefact     <- runBoot bootTracer (pure . Macro.cExpr) config [header]
+      bootArtefact     <- runBoot bootTracer (pure . Macro.cExpr) config
+                            [C.DirectiveHashInclude header]
       frontendArtefact <- runFrontend frontendTracer frontendConfig bootArtefact
       k frontendArtefact
   where